==> 02-Augment.ipynb <==
# Augment the Data

Let us augment the items(movie) data with more fields from IMDB and / or TMDB data

import numpy as np
import pandas as pd
import json, glob
import requests
import re
import tmdb
from importlib import reload
from reco.recoutils import create_directory, update_progress

## Get the Movie Features from TMDB

- We will use the "links.csv" which contains:
    - movie_id
    - imdb_id
    - tmdb_id
- Source: Scraper + Manual Enhancement
     - Bootstrapped using the scraper to get the IMDB title link
     - Enhanced through search on TMDB database to get complete list

links = pd.read_csv("links.csv", dtype={"tmdb_id": 'Int64' })

links.head()

links.shape

## Get Movie Feature from TMDB 

# sample_id = links.tmdb_id[5]
# tmdb.movie(sample_id)

tmdb_ids = links.tmdb_id.tolist()
print(tmdb_ids)

def get_movie_features(tmdb_ids):
    num = len(tmdb_ids)
    tick = 0
    features = []
    for i in tmdb_ids:
        feature = tmdb.movie(i)
        if feature != False: 
            features.append(feature)
        
        tick = tick + 1
        update_progress(tick / num)
    
    update_progress(1)
    return features

def save_movie_features(features):
    outfile = "data/features.json"
    with open(outfile, 'w') as fout:
            json.dump(features, fout)

# Uncomment and Run this to download movie features from tmdb

#features = get_movie_features(tmdb_ids)
#save_movie_features(features)

## Create Item Features Augment

df_features = pd.read_json("data/features.json")
#df_features = pd.DataFrame.from_dict(features)

# Select the non NA links
links_full = links[~links.tmdb_id.isna()][["movie_id", "tmdb_id"]]

df_item_features = pd.merge(left=df_features, right=links_full, left_on="id", right_on="tmdb_id", how="left")

df_item_features.head()

df_item_features.to_csv("data/item_features.csv", index=None)

## Get Movie Posters

# Create Poster Directory
create_directory("/data/posters")

def poster(movie_id, df):
    """
    Gets the movie poster from the movie_id
    
    poster_path (string): path for the poster
    df (pd.Dataframe): pandas dataframe with movie id and poster_path
    """
    directory_path = "data/posters/"    
    poster_path = df[df.movie_id == movie_id].poster_path.values[0]
    if poster_path == None:
        poster_path = ""
    
    save_path = directory_path + str(movie_id) + ".jpg"
    
    base_url = "https://image.tmdb.org/t/p/w92"
    url = base_url + poster_path   
    
    response = requests.request("GET", url, stream=True)
    if response.status_code == 200:
        with open(save_path, 'wb') as out_file:
            response.raw.decode_content = True 
            shutil.copyfileobj(response.raw, out_file)
        return True
    else: 
        return False

df_item_features[df_item_features.poster_path.isna()][["movie_id", "poster_path"]]

import shutil

#poster(1, df_item_features)

#from IPython.display import Image 
#Image(filename='data/posters/8.jpg')

def get_all_posters(df):
    """
    Get all posters from the dataframe and save with movie_id
    """
    id_list = df.movie_id.tolist()
    num = len(id_list)
    tick = 0
    
    for i in id_list:
        df["image"] = poster(i, df)
        tick = tick + 1
        update_progress(tick / num)
    
    update_progress(1)
    return df

#Uncomment and run this to get posters
#get_all_posters(df_item_features)
