==> 03-Refine.ipynb <==
# Refine the Data

Create the items, users and interaction matrix

- **Items**: item id + metadata features
- **Users**: User id + metadata features
- **Interaction**: Explicit Matrix, Implicit Matrix

import numpy as np
import pandas as pd
import re
import os

## Items

For the items, we will keep the following:
- From `items_raw` dataframe
  - movie_id
  - title
  - year (release)
  - genre categories
- From `item_features` dataframe
  - overview
  - language (original_language)
  - runtime
  - vote_average
  - vote_count

items_raw = pd.read_csv("data/items_raw.csv")
item_features = pd.read_csv("data/item_features.csv")

*1. Get `year` from `release_title`* 

items_raw["release_date"] = pd.to_datetime(items_raw.release_date, infer_datetime_format=True)
items_raw["year"] = items_raw.release_date.apply(lambda x: str(x.year))

*2. Drop `imdb_url`,  `video_release_date` & `release_date`*

items_main = items_raw.drop(['video_release_date', 'release_date', 'imdb_url'], axis=1).copy()

# Match Whitespace + ( + YEAR + )
# regex_year = re.compile(r'\s\(\d{4}\)')
# items["movie"] = items.title.str.replace(regex_year, "")
# items["movie"] = items.movie.str.strip()

*3. Get the additional features from the item_features*

items_addtl = item_features[['overview', 'original_language', 'runtime', 'vote_average', 'vote_count', "movie_id"]].copy()

*4. Merge the two dataframes*

items = pd.merge(left=items_main, right=items_addtl, on="movie_id", how="left")

items.head()

items.to_csv("data/items.csv", index=None)

# Function to remove the year from the title
def remove_year(title):
    import re
    # Remove the year in parentheses at the end of the title
    return re.sub(r'\s\(\d{4}\)$', '', title)

df_items = pd.read_csv("data/items.csv")

# Apply the function to the 'title' column
df_items['cleaned_title'] = df_items['title'].apply(remove_year)

# Display the updated DataFrame
print(df_items[['movie_id', 'title', 'cleaned_title']].head())

# def remove_trailing_the(title):
#     if title.lower().endswith(', the'):
#         return title[:-5].strip()
#     return title

#df_items['cleaned_title'] = df_items['cleaned_title'].apply(remove_trailing_the)

#df_items.to_csv("data/items.csv", index=False)

# Users

import spacy
from spacy.lang.en.examples import sentences 

nlp = spacy.load('en_core_web_lg')
doc = nlp(sentences[0])
print(doc.text)
for token in doc:
    print(token.text, token.pos_, token.dep_)

import spacy

#nlp = spacy.load('en_core_web_md')  # make sure to use larger model!
tokens = nlp(u'dog cat banana')

for token1 in tokens:
    for token2 in tokens:
        print(token1.text, token2.text, token1.similarity(token2))

# import umap
# from sklearn.datasets import load_digits

# digits = load_digits()

# embedding = umap.UMAP().fit_transform(digits.data)
