==> 04-Transform.ipynb <==
# Transform

We need to prepare the data for using Deep Learning algorithms. There are two main transformations:

- **Preprocessing**:
    - Users: Convert to label encoded values
    - Items: Convert to label encoded values
    - User Side Features: Convert to 
    - Ratings: Convert the interaction to Explicit or Implicit signals


- **Data Spliting**: Create train & test datasets for evaluating the dataset
    - Random Split
    - Stratified Split
    - Chronological Split
    

# set the environment path to find reco
import sys
sys.path.append("../")

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

users = pd.read_csv("data/users.csv")
items = pd.read_csv("data/items.csv")
ratings = pd.read_csv("data/ratings.csv")

users.head()

ratings.head()

items.head()

items.iloc[45]

## Preprocessing

from reco.preprocess import encode_user_item_withencoder

encoded_ratings, user_encoder, item_encoder = encode_user_item_withencoder(ratings, "user_id", "movie_id", "rating", "unix_timestamp")

encoded_ratings.head()

ratings['rating'] = ratings['rating'].values.astype(np.float32)
min_rating = min(ratings['rating'])
max_rating = max(ratings['rating'])
print(min_rating, max_rating)

## Data Splitting

- Random Split
- Stratified Split
- Chronological Split

from reco.preprocess import random_split, user_split, sample_data

sampledf = sample_data()
print(sampledf)

### Random Split 

train_random, val_random, test_random = random_split(sampledf, [0.6, 0.2, 0.2])

train_random

val_random

test_random

### User Chronological Split

train_chrono, val_chrono, test_chrono = user_split(sampledf, [0.6, 0.2, 0.2])
#print(sampledf)

train_chrono

val_chrono

test_chrono
