==> 05-Evaluation.ipynb <==
# Evaluation

## Prediction Metrics (Similiar to Regression Problem)
    - RMSE
    - MSE
    - MAE

### Hit Metrics (Similiar to Classification Metrics)
**Hit** - defined by relevancy, a hit usually means whether the recommended "k" items hit the "relevant" items by the user. For example, a user may have clicked, viewed, or purchased an item for many times, and a hit in the recommended items indicate that the recommender performs well. Metrics like "precision", "recall", etc. measure the performance of such hitting accuracy.

    - Precision@k
    - Recall@k
  

### Ranking Metrics

**Ranking** - ranking metrics give more explanations about, for the hitted items, whether they are ranked in a way that is preferred by the users whom the items will be recommended to. Metrics like "mean average precision", "ndcg", etc., evaluate whether the relevant items are ranked higher than the less-relevant or irrelevant items. 

    - MeanReciprocalRank@k
    - MeanAveragePrecision@k
    - NDCG@k


import numpy as np
import pandas as pd

df_true = pd.DataFrame(
        {
            "USER": [1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3],
            "ITEM": [1, 2, 3, 1, 4, 5, 6, 7, 2, 5, 6, 8, 9, 10, 11, 12, 13, 14],
            "RATING": [5, 4, 3, 5, 5, 3, 3, 1, 5, 5, 5, 4, 4, 3, 3, 3, 2, 1],
        }
    )

df_pred = pd.DataFrame(
    {
        "USER": [1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3],
        "ITEM": [3, 10, 12, 10, 3, 11, 5, 13, 4, 10, 7, 13, 1, 3, 5, 2, 11, 14],
        "RATING": [14, 13, 12, 14, 13, 12, 11, 10, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5]
    }
)

from recoflow.metrics import MeanAbsoluteError, MeanSquaredError, RootMeanSquaredError
from recoflow.metrics import PrecisionK, RecallK, NDCGK, MeanReciprocalRankK, MeanAveragePrecisionK

RootMeanSquaredError(df_true, df_pred)

MeanSquaredError(df_true, df_pred)

MeanAbsoluteError(df_true, df_pred)

PrecisionK(df_true, df_pred, 3)

RecallK(df_true, df_pred, 3)

NDCGK(df_true, df_pred, 3)

MeanAveragePrecisionK(df_true, df_pred, 3)

MeanReciprocalRankK(df_true, df_pred, 3)

rating_true = df_true
rating_pred = df_pred
k = 5

common_users = set(rating_true["USER"]).intersection(set(rating_pred["USER"]))
rating_true_common = rating_true[rating_true["USER"].isin(common_users)]
rating_pred_common = rating_pred[rating_pred["USER"].isin(common_users)]
n_users = len(common_users)

rating_pred_common.head(8)

from recoflow.metrics import _GetTopKItems

df_hit = _GetTopKItems(rating_pred_common, "USER", "RATING", k)

rating_true_common.head(6)

df_hit.head(6)

df_hit = pd.merge(df_hit, rating_true_common, on=["USER", "ITEM"])[
        ["USER", "ITEM", "rank"]]

df_hit

# count the number of hits vs actual relevant items per user
df_hit_count = pd.merge(
    df_hit.groupby("USER", as_index=False)["USER"].agg({"hit": "count"}),
    rating_true_common.groupby("USER", as_index=False)["USER"].agg(
        {"actual": "count"}
    ),
    on="USER",
)


df_hit_count

#return df_hit, df_hit_count, n_users
